core: unify srUtilItoA calling sequence
This helps against improper propagation.

+++ runtime/stream.c
@@ -1916,7 +1916,7 @@ static rsRetVal strmWriteLong(strm_t *__restrict__ const pThis, const long i) {
 
     assert(pThis != NULL);
 
-    CHKiRet(srUtilItoA((char *)szBuf, sizeof(szBuf), i));
+    CHKiRet(srUtilItoA((char *)szBuf, sizeof(szBuf), (int64_t) i));
     CHKiRet(strmWrite(pThis, szBuf, strlen((char *)szBuf)));
 
 finalize_it:
